refactor: keep production modules within release limits
